Fix Nashorn issue with map lookup

+++ molgenis-js/src/test/java/org/molgenis/js/JsMagmaScriptEvaluatorTest.java
@@ -547,7 +547,7 @@ public class JsMagmaScriptEvaluatorTest {
     Entity entity0 = new DynamicEntity(personHeightEntityType);
     entity0.set("height", 180);
     Object result = jsMagmaScriptEvaluator.eval("$('height').plus(100).value()", entity0, 3);
-    assertEquals(result, (double) 280);
+    assertEquals(result, 280d);
   }
 
   @Test
@@ -556,7 +556,7 @@ public class JsMagmaScriptEvaluatorTest {
     entity0.set("height", 180);
     Object result1 =
         jsMagmaScriptEvaluator.eval("$('height').plus(new newValue(100)).value()", entity0, 3);
-    assertEquals(result1, (double) 280);
+    assertEquals(result1, 280);
   }
 
   @Test
@@ -572,7 +572,7 @@ public class JsMagmaScriptEvaluatorTest {
     Entity entity0 = new DynamicEntity(personHeightEntityType);
     entity0.set("height", 2);
     Object result = jsMagmaScriptEvaluator.eval("$('height').times(100).value()", entity0, 3);
-    assertEquals(result, (double) 200);
+    assertEquals(result, 200d);
   }
 
   @Test
@@ -588,7 +588,7 @@ public class JsMagmaScriptEvaluatorTest {
     Entity entity0 = new DynamicEntity(personHeightEntityType);
     entity0.set("height", 20);
     Object result = jsMagmaScriptEvaluator.eval("$('height').pow(2).value()", entity0, 3);
-    assertEquals(result, 400d);
+    assertEquals(result, 400);
   }
 
   @Test
@@ -631,7 +631,7 @@ public class JsMagmaScriptEvaluatorTest {
     Entity person = new DynamicEntity(personBirthDateMeta);
     person.set("birthdate", now().atOffset(UTC).toLocalDate());
     Object result = jsMagmaScriptEvaluator.eval("$('birthdate').age().value()", person);
-    assertEquals(result, 0d);
+    assertEquals(result, 0);
   }
 
   @Test
@@ -643,7 +643,7 @@ public class JsMagmaScriptEvaluatorTest {
     Collection<Object> result =
         jsMagmaScriptEvaluator.eval(
             Arrays.asList("$('weight').value()", "$('height').pow(2).value()"), person);
-    assertEquals(result, Arrays.asList(80, 400d));
+    assertEquals(result, Arrays.asList(80d, 400d));
   }
 
   @Test(enabled = false)
